Moving context form the pkg to cmd.
This move is to make sure:
* Pkg code is independent of the context (which is very pf9ctl
specific).
* Hide encoding decoding of password to the context code only
* Fix other bugs
