+++ cmd/root.go
@@ -7,7 +7,6 @@ import (
 	"os"
 
 	homedir "github.com/mitchellh/go-homedir"
-	"github.com/platform9/pf9ctl/pkg/constants"
 	"go.uber.org/zap"
 	"github.com/platform9/pf9ctl/pkg/log"
 	"github.com/spf13/cobra"
@@ -33,7 +32,7 @@ func Execute() {
 	}
 
 	// Initializing zap log with console and file logging support
-	if err := log.ConfigureGlobalLog(true, constants.Pf9Log); err != nil {
+	if err := log.ConfigureGlobalLog(false, Pf9Log); err != nil {
 		fmt.Printf("log initialization failed: %s", err.Error())
 		os.Exit(1)
 	}
@@ -44,15 +43,15 @@ func Execute() {
 }
 
 func initializeBaseDirs() (err error) {
-	err = os.MkdirAll(constants.Pf9Dir, 0700)
+	err = os.MkdirAll(Pf9Dir, 0700)
 	if err != nil {
 		return 
 	}
-	err = os.MkdirAll(constants.Pf9DBDir, 0700)
+	err = os.MkdirAll(Pf9DBDir, 0700)
 	if err != nil {
 		return
 	}
-	err = os.MkdirAll(constants.Pf9LogDir, 0700)
+	err = os.MkdirAll(Pf9LogDir, 0700)
 	return
 }
