--- pkg/util/keystone/keystone.go
@@ -1,1 +1,0 @@
-package keystone
